+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Switch from `centos/7` to `ubuntu/focal64` vagrant box
 - Add ARM64 build and test script for vagrant
 - Add ARM64 build for Github Actions
+- Add coverage report
 
 ## [0.1.2] - 2022-03-07